+++ examples/data-sources/nxos_pim_interface/data-source.tf
@@ -1,0 +1,4 @@
+data "nxos_pim_interface" "example" {
+  name         = "default"
+  interface_id = "eth1/10"
+}