+++ examples/resources/nxos_pim_interface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_pim_interface.example "sys/pim/inst/dom-[default]/if-[eth1/10]"