+++ examples/resources/nxos_pim_interface/resource.tf
@@ -1,0 +1,9 @@
+resource "nxos_pim_interface" "example" {
+  name         = "default"
+  interface_id = "eth1/10"
+  admin_state  = "enabled"
+  bfd          = "enabled"
+  dr_priority  = 10
+  passive      = false
+  sparse_mode  = true
+}